KEYCLOAK-12265 Move KerberosEmbeddedServer to testsuite
