+++ examples/kerberos/pom.xml
@@ -96,7 +96,7 @@
                         <groupId>org.codehaus.mojo</groupId>
                         <artifactId>exec-maven-plugin</artifactId>
                         <configuration>
-                            <mainClass>org.keycloak.util.ldap.KerberosEmbeddedServer</mainClass>
+                            <mainClass>org.keycloak.testsuite.KerberosEmbeddedServer</mainClass>
                             <classpathScope>test</classpathScope>
                             <systemProperties>
                                 <systemProperty>
