+++ testsuite/integration-arquillian/tests/base/src/main/java/org/keycloak/testsuite/util/KerberosRule.java
@@ -23,7 +23,7 @@ import java.util.Properties;
 
 import org.jboss.logging.Logger;
 import org.keycloak.testsuite.client.resources.TestingResource;
-import org.keycloak.util.ldap.KerberosEmbeddedServer;
+import org.keycloak.testsuite.KerberosEmbeddedServer;
 import org.keycloak.util.ldap.LDAPEmbeddedServer;
 
 /**
