+++ testsuite/utils/src/main/java/org/keycloak/testsuite/KerberosEmbeddedServer.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.keycloak.util.ldap;
+package org.keycloak.testsuite;
 
 import org.apache.directory.api.ldap.model.constants.SupportedSaslMechanisms;
 import org.apache.directory.api.ldap.model.exception.LdapInvalidDnException;
@@ -35,6 +35,7 @@ import org.apache.directory.shared.kerberos.KerberosTime;
 import org.apache.directory.shared.kerberos.KerberosUtils;
 import org.apache.directory.shared.kerberos.codec.types.EncryptionType;
 import org.jboss.logging.Logger;
+import org.keycloak.util.ldap.LDAPEmbeddedServer;
 
 import javax.security.auth.kerberos.KerberosPrincipal;
 import java.io.IOException;
